+++ src/Services/Dominion/API/InvadeCalculationService.php
@@ -164,7 +164,13 @@ class InvadeCalculationService
         $returningForcesDP = $this->militaryCalculator->getDefensivePower($dominion, null, null, $unitsReturning, 0, true);
         $homeForcesDP = $this->militaryCalculator->getDefensivePower($dominion);
 
-        $this->calculationResult['max_op'] = $this->calculationResult['home_defense'] * 1.25;
+        $ratio = 1.25;
+        $recentInvasions = $this->militaryCalculator->getRecentlyInvadedCount($dominion, 5 * 24, true);
+        if ($recentInvasions > 2) {
+            $ratio = 1.25 - min(0.5, ($recentInvasions - 2) * 0.125);
+        }
+
+        $this->calculationResult['max_op'] = $this->calculationResult['home_defense'] * $ratio;
         $this->calculationResult['min_dp'] = ($returningForcesDP + $homeForcesDP) * 0.40;
 
         $this->calculationResult['target_min_dp'] = $this->militaryCalculator->getMinimumDefense($target);